extracted Tokens to separate file and used in gocode
